exercise ex01

+++ 02_variables_data_types/ex01/solution.c
@@ -6,7 +6,7 @@
 /*   By: shobeedev <https://shobee.space/>                   / __/   ___) |   */
 /*                                                          |_____| |____/    */
 /*   Created: 2026/03/10 11:02:05 by shobeedev               shobee4ever      */
-/*   Updated: 2026/03/10 11:08:36 by shobeedev            tfaaty fi l3oolaa   */
+/*   Updated: 2026/03/10 11:17:00 by shobeedev            tfaaty fi l3oolaa   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,11 +22,11 @@ int main()
 	int sum = a + b;
 	int diff = a - b;
 	int prod = a * b;
-	float quot = a / b;
+	int quot = a / b;
 	printf("%d + %d = %d\n",a,b,sum);
 	printf("%d - %d = %d\n",a,b,diff);
 	printf("%d * %d = %d\n",a,b,prod);
-	printf("%d / %d = %.2f\n",a,b,quot);
+	printf("%d / %d = %d\n",a,b,quot);
 
 	return 0;
 }
